project2/requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
flask
pyfaktory
python-dotenv
matplotlib
plotly
